[fixes #151]
The basic.sh tests work, but the script is not pretty. I plan to add
more tests within the next few days.

My hack to fix the make_dir function in utils_rmw.c should be okay for a
while. The only time I've had a problem with it is when I used a "." for
a HOME directory.

I made sure the make_dir patch worked before I changed HOME to tmp-home
in VARS