write a default config file if one isn't present
Thanks again for the feedback, @alketii

Also included with this patch, if verbosity is on, tell the user which
config file is being read.

This may help with debugging as well. I had a problem last week when I
was getting unexpected results when working on a new patch. The problem
was that rmw was reading the wrong config file. It took me 2 hours to
realize it.